refactor: refactor parser utilities and unify soupify usage; improve number extraction
